Moved sensitive values to env and added Google Login with User model
Sensitive values were hardcoded so I have moved to an env file. Find the contents of the env file in onedrive. Google login now working and finding user in database. Need to create user when doesn't exist in db. Still need to handle when to show auth and when not.