Hooked up creating a round from a chosen course and holes
Once a user is logged in, they can choose to start a round from the
menu. The menu is now mobile friendly. In order to start a round a
course must be chosen. You are able to see a list of holes and a full
scorecard. Next to allow scoring for each hole.