Close sidebar on nav and refactor auth/user
Before a route is resolved it now closes the drawer. Next, auth happened
in the Home view so I have moved it to the store. This still needs
review. I also need to work out what is going on with the select element
and hole 1. Then we need scores that are saved to save into the
scorecard tab.