Add dark mode and setup modal form
Changed color scheme to use dark mode. This should help save on battery
life and I think it looks better. We'll see what users say. ScoreForm is
now in a modal that is activated by the score td. Need to work on
getting the scores related to the appropriate hole.