Added ability to close round
I tested the application during a round and it went well. Realized we
didn't have a way to show total score so that was added to each 9. Also
didn't have a way to end a round in progress. Added a button to the
bottom of the scorecard to close the round.