Added close round and Hilldale
I have added the ability to close a round so you can move on to another
round. Kinda important. I've also added a seed for Hilldale, but removed
the sync call so that the data isnt removed from teh db. I have  a
backup of the db just in case.