Redesigning round workflow
I've begun work on the workflow of starting a new round. This will now
be in steps to support selecting options as well as inviting other
users. This is part of a large new feature I want to add; inviting
others to your round. This can help expedite creating rounds for others
as well as make scores visible to the group.