Closing round and editing scores works!
I have finished the work to send a score from the Score Form as well as
close out the round for the user. Currently there is no way to retrieve
a closed round so I will need to add that in. Maybe in the form of a
'recent rounds' section?