Add dynamic stats to Stats tab
Putts and round average is currently working. Putts only count against
holes that users enter a value for. This is very much in the hands of
the user. Handicap is setup to say N/A if the user has less than 8
games. Will hook this up to correctly calculate handicap in future.